+++ src/icemac/addressbook/browser/search/result/handler/test_delete.py
@@ -60,6 +60,7 @@ def test_delete__DeleteForm__3(search_data, browser, role):
          'E-Mail (Creates a link to send e-mails.)',
          'Names (Comma separated list of person names.)',
          'Checklist (List of person names with check-boxes.)',
+         'Birthday list (Person names sorted by birthday.)',
          "iCalendar export birthday (Export person's birthdays as "
          ".ics file.)"] ==
         browser.getControl('Apply on selected persons').displayOptions)